+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/components/gesturehandler/common/GestureHandlerStateManager.kt
@@ -1,5 +1,0 @@
-package versioned.host.exp.exponent.modules.api.components.gesturehandler
-
-interface GestureHandlerStateManager {
-  fun setGestureHandlerState(handlerTag: Int, newState: Int)
-}