+++ ios/vendored/unversioned/react-native-gesture-handler/ios/RNGestureHandlerButtonComponentView.mm
@@ -3,13 +3,13 @@
 #import "RNGestureHandlerButtonComponentView.h"
 
 #import <React/RCTConversions.h>
+#import <React/RCTFabricComponentsPlugins.h>
 
-#import <react/renderer/components/rngesturehandler/ComponentDescriptors.h>
-#import <react/renderer/components/rngesturehandler/EventEmitters.h>
-#import <react/renderer/components/rngesturehandler/Props.h>
-#import <react/renderer/components/rngesturehandler/RCTComponentViewHelpers.h>
+#import <react/renderer/components/rngesturehandler_codegen/ComponentDescriptors.h>
+#import <react/renderer/components/rngesturehandler_codegen/EventEmitters.h>
+#import <react/renderer/components/rngesturehandler_codegen/Props.h>
+#import <react/renderer/components/rngesturehandler_codegen/RCTComponentViewHelpers.h>
 
-#import "RCTFabricComponentsPlugins.h"
 #import "RNGestureHandlerButton.h"
 
 using namespace facebook::react;